Patch from Debian bug #74391 to compile cleanly on Alpha.

+++ emit.h
@@ -41,7 +41,7 @@ int             _hs_emit_filesum(hs_write_fn_t write_fn, void *write_priv,
 int             _hs_emit_literal_cmd(hs_write_fn_t write_fn, void *write_priv,
 				     size_t size);
 
-int             _hs_emit_checksum_cmd(hs_write_fn_t, void *, uint32_t size);
+int             _hs_emit_checksum_cmd(hs_write_fn_t, void *, size_t size);
 
 int             _hs_emit_copy(hs_write_fn_t write_fn, void *write_priv,
 			      off_t offset, size_t length,
